backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.5
flask-cors==4.0.0
python-engineio==4.5.1
python-socketio==5.8.0
eventlet==0.33.3
python-dotenv
